Create ROFL Orb Project Sem_2.cpp
This project was made by us in Semester 2 in Software Development Fundamental subject with all the aquired knowledge in C++. Topics that were used to make this project were  file handling, pointers, data structures, and a topic that we learnt on our own  that is graphics. This project was a game in which a ball will come down from the top and thier are slabs coming from bottom using which the ball can stay on the screen and eat the food that are alotted randomly anywhere on the screen. Score will depend on the number of food item eated by the ball and the ball will have 10 lives after which its score gets stored in the file.